+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-14T13:06:01.809Z",
+    "2026-07-14T14:40:08.432Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {